:children_crossing: update body Specs in Nutrition Screen

+++ lib/domain/entities/nutrition_entities/body_specs.dart
@@ -67,8 +67,8 @@ class Pal extends Equatable {
 }
 
 class Bmr extends Equatable {
-  final int? bmrPerKg;
-  final int? bmrPerDay;
+  final double? bmrPerKg;
+  final double? bmrPerDay;
 
   Bmr({
     this.bmrPerKg,
